password check before deleting

+++ user/user.delete.js
@@ -1,4 +1,5 @@
 import { object, string, number } from 'yup';
+import { compare } from 'bcrypt';
 import connectionPool from '../_utilities/connection.js';
 import APIError from '../_utilities/apiError.js';
 
@@ -14,8 +15,14 @@ async function deleteUser(user) {
     user = await userSchema.validate(user);
 
     // TODO - This needs to cascade across all data that has user ID and use a trasaction
-    const sqlquery = `DELETE FROM users WHERE user_id = ? AND username = ?`;
-    const dbResponse = (await connectionPool.execute(sqlquery, [user.user_id, user.username]));
+    // Double check auth and password before deleting
+    const sqlUserquery = `SELECT * FROM users WHERE user_id = ? AND username = ?`;
+    let dbResponse = (await connectionPool.execute(sqlUserquery, [user.user_id, user.username]))[0][0];
+    if (!dbResponse) throw new APIError('Database connection Error', 500, 'Unable to get response from Database');
+    if (!(await compare(user.password, dbResponse.password))) throw new APIError('Bad Auth', 401, 'Password is incorrect');
+
+    const sqlDeletequery = `DELETE FROM users WHERE user_id = ? AND username = ?`;
+    dbResponse = (await connectionPool.execute(sqlDeletequery, [user.user_id, user.username]));
     if (dbResponse.affectedRows == 0) throw new APIError('Reject Data', 400, 'Failed to find user');
 
     return {
